Migrate experiments and helpfulness to Hydro (#16059)
* Migrate experiments and helpfulness to Hydro

* Clean out old tests

* ...and more old tests to delete
