+++ javascripts/experiment.js
@@ -1,47 +1,21 @@
 import murmur from 'imurmurhash'
-import { v4 as uuidv4 } from 'uuid'
-import Cookies from 'js-cookie'
-import getCsrf from './get-csrf'
+import { getUserEventsId, sendEvent } from './events'
 
 const TREATMENT = 'TREATMENT'
 const CONTROL = 'CONTROL'
-const COOKIE_NAME = '_docs-experiment'
-
-let cookieValue
-
-export function getUserExperimentId () {
-  if (cookieValue) return cookieValue
-  cookieValue = Cookies.get(COOKIE_NAME)
-  if (cookieValue) return cookieValue
-  cookieValue = uuidv4()
-  Cookies.set(COOKIE_NAME, cookieValue, {
-    secure: true,
-    sameSite: 'strict',
-    expires: 365
-  })
-  return cookieValue
-}
 
 export function bucket (test) {
-  const id = getUserExperimentId()
+  const id = getUserEventsId()
   const hash = murmur(test).hash(id).result()
   return hash % 2 ? TREATMENT : CONTROL
 }
 
 export async function sendSuccess (test) {
-  return fetch('/events', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      'CSRF-Token': getCsrf()
-    },
-    body: JSON.stringify({
-      type: 'EXPERIMENT',
-      user: getUserExperimentId(),
-      test,
-      group: bucket(test).toLowerCase(),
-      success: 'yes'
-    })
+  return sendEvent({
+    type: 'experiment',
+    experiment_name: test,
+    experiment_variation: bucket(test).toLowerCase(),
+    experiment_success: true
   })
 }
